feat(phy): begin soft-decision LLR pipeline
Change phy_ip buffer type from uint8_t to float to carry LLR values
across PHY→IP boundary.

- Add descramble_llr(): applies LFSR descrambling directly in LLR
  domain by negating values where LFSR feedback bit is 1
- Switch run_dsp_rx to call descramble_llr(llr) and write LLR to
  phy_ip instead of hard-decision bits
- Add temporary hard-decision stub in run_tun_rx (sign of LLR → bit)
  until full soft-decision decoding is implemented

+++ include/common.hpp
@@ -246,7 +246,7 @@ struct SharedData {
     DSP dsp;
 
     DoubleBuffer<uint8_t> ip_phy;
-    DoubleBuffer<uint8_t> phy_ip;
+    DoubleBuffer<float> phy_ip;
 
     DoubleBuffer<int16_t> sdr_dsp_tx;
     DoubleBuffer<int16_t> sdr_dsp_rx;
